Relationship between age, gender, and SATV score
Through R, I was able to analyse the relationship between age, gender, and, in this example, SATV scores. While SATV scores and the specific packages used for this analysis—library(psych) and sat.act—are primarily associated with test scores for college admissions in the United States, measuring skills such as writing, critical reading, and mathematics, they offered me an insightful first glimpse into how gender and age intersect in learning and performance. 
These visualizations revealed trends that would have been difficult to perceive without graphical representation, providing clarity and context to abstract data. Importantly, they also highlighted how steep the learning curve can be—not only for coding but for learning in general—and how this might be influenced by age.
It is thanks to this incredible potential to explore and visualize data that R has become more than just a tool; it has become a lens through which I can better understand the world, my own learning journey, and the factors that shape it.